requirements.txt
Flask==2.0.2
Flask_Bcrypt==1.0.1
Flask_Login==0.6.2
Flask_Reuploaded==1.2.0
Flask_SQLAlchemy==2.5.1
flask_uploads==0.2.1
Flask_WTF==1.0.1
matplotlib==3.5.1
numpy==1.20.0
opencv_python==4.5.5.62
Pillow==9.3.0
tensorflow==2.7.0
tensorflow_hub==0.12.0
WTForms==3.0.1
